tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.size.attributes.parse.empty.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.canvas.host.size.attributes.parse.empty</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.size.attributes.parse.empty</h1>
<p class="desc">Parsing of non-negative integers</p>


<script>
var t = async_test("Parsing of non-negative integers");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  canvas.width = '';
  canvas.height = '';
  _assertSame(canvas.width, 0, "canvas.width", "0");
  _assertSame(canvas.height, 0, "canvas.height", "0");
  t.done();

});
</script>
